........
fixed gcc errors
........
- merged issue 1450
........
merged 1435
........
work in proress... MaskBoost added...
........
BoostMaskAccessor added. Work in progress...
........

........
code/io/mask added, ITKImageMaskAccessor implemented.
........
ITKImageMaskAccessor generator classes added.
........
Testing/io/mask added
........
ITKImageMaskAccessorConverter added. Test added.
........
minor modification
........
Class name modified. BoostMaskAccessorTest added.
........
Test added.
........
delete debug codes
........
delete debug codes
........
redundant codes in rttbITKImageFileDoseAccessorGenerator and rttbITKImageFileMaskAccessorGenerator moved to rttbITKIOHelper
........
updated based on review comments: rttbBoostMask voxelization steps moved from constructor to calcMask() 
........

+++ testing/core/DummyDVHGenerator.cpp
@@ -22,63 +22,87 @@
 #include <time.h>
 
 #include "DummyDVHGenerator.h"
-namespace rttb{
-	namespace testing{
+namespace rttb
+{
+	namespace testing
+	{
 
 		DummyDVHGenerator::DummyDVHGenerator(): _binSize(DoseTypeGy(0.1)), _voxelVolume(8), _value(0)
 		{
 			/* initialize random seed: */
-			srand (time(NULL));
+			srand(static_cast<unsigned int>(time(NULL)));
 		};
 
-		core::DVH DummyDVHGenerator::generateDVH(IDType structureID,IDType doseID){				
+		core::DVH DummyDVHGenerator::generateDVH(IDType structureID, IDType doseID)
+		{
 			core::DVH::DataDifferentialType aDataDifferential;
-			for(int i = 0; i < 100; i++){
-				_value = DoseCalcType((double(rand())/RAND_MAX)*1000);
+
+			for (int i = 0; i < 100; i++)
+			{
+				_value = DoseCalcType((double(rand()) / RAND_MAX) * 1000);
 				//cut off values to avoid problems on comparisson with reimported values after
 				//writing to file.
-				_value = floor(_value *1000000)/1000000;
-				aDataDifferential.push_back( _value );
+				_value = floor(_value * 1000000) / 1000000;
+				aDataDifferential.push_back(_value);
 				}
+
 			return core::DVH(aDataDifferential, _binSize, _voxelVolume, structureID, doseID);
 			}
 
-		core::DVH DummyDVHGenerator::generateDVH(IDType structureID,IDType doseID, DoseCalcType value){
+		core::DVH DummyDVHGenerator::generateDVH(IDType structureID, IDType doseID, DoseCalcType value)
+		{
 			_value = value;
-			
+
 			core::DVH::DataDifferentialType aDataDifferential;
-			for(int i = 0; i < 100; i++){
-				aDataDifferential.push_back( _value );
+
+			for (int i = 0; i < 100; i++)
+			{
+				aDataDifferential.push_back(_value);
 		}
+
 			return core::DVH(aDataDifferential, _binSize, _voxelVolume, structureID, doseID);
 		}
 
-		core::DVH DummyDVHGenerator::generateDVH(IDType structureID,IDType doseID, DoseCalcType minValue, DoseCalcType maxValue){
-			_voxelVolume = 0.2*0.2*0.4;
+		core::DVH DummyDVHGenerator::generateDVH(IDType structureID, IDType doseID, DoseCalcType minValue,
+		        DoseCalcType maxValue)
+		{
+			_voxelVolume = 0.2 * 0.2 * 0.4;
 			bool decrease = false;
-			
+
 			core::DVH::DataDifferentialType aDataDifferential;
-			for(int i = 0; i <= 200; i++){
-				if ((i > 20) && (i < 180)){
-					if ((_value == 0) && (!decrease)) {
-						_value = DoseCalcType((double(rand())/RAND_MAX)*10)+minValue;
+
+			for (int i = 0; i <= 200; i++)
+			{
+				if ((i > 20) && (i < 180))
+				{
+					if ((_value == 0) && (!decrease))
+					{
+						_value = DoseCalcType((double(rand()) / RAND_MAX) * 10) + minValue;
 					}
-					else if (!decrease){
-						_value = _value + DoseCalcType((double(rand())/RAND_MAX)*(maxValue/10));
+					else if (!decrease)
+					{
+						_value = _value + DoseCalcType((double(rand()) / RAND_MAX) * (maxValue / 10));
 					}
-					if ( (_value > maxValue) || (decrease) ){
+
+					if ((_value > maxValue) || (decrease))
+					{
 						decrease = true;
-						_value = _value - DoseCalcType((double(rand())/RAND_MAX)*(maxValue/3));
+						_value = _value - DoseCalcType((double(rand()) / RAND_MAX) * (maxValue / 3));
 					}
-					if (_value < 0) {
+
+					if (_value < 0)
+					{
 						_value = 0;
 					}
 				}
-				else{
+				else
+				{
 					_value = 0;
 				}
-				aDataDifferential.push_back( _value );
+
+				aDataDifferential.push_back(_value);
 			}
+
 			return core::DVH(aDataDifferential, _binSize, _voxelVolume, structureID, doseID);
 		}
 	}
